requirements.txt
# Data Collection & Processing
pandas
numpy
requests
python-dotenv
yfinance
fredapi
lxml

# Machine Learning
scikit-learn
xgboost
optuna
scipy

# Visualization
matplotlib
seaborn

# Jupyter
jupyter
ipykernel

# Utilities
python-dateutil